Do not allow help for command roles other than main.
Previously requesting help for roles other than main would result in a segfault since help.auto.c.inc does not include such help.

Prevent this by erroring when a user requests help for roles other than main.

This is a bug but it is unlikely to have ever been seen in the field and in any case it has a very low severity since it only impacts a corner-case for the help command.