Fix postgres/client throttled by the wait handler on large results.
The asynchronous logic used to implement the query timeout is misusing `PQisBusy()`, which causes the wait handler throttling the consumption of command results. It can introduce a huge delay on a backup job up to `db-timeout` because of the back-off sequence. `PQconsumeInput()` appears to be reading 16KB at a time.

Following the recommendation of libpq, the fix is to watch the client socket for data availability and keep consuming results and checking for command busyness.

Signed-off-by: Thibault VINCENT <thibault.vincent@enix.fr>