Enable missing-prototypes compiler warning.
Warn if a global function is defined without a previous prototype declaration. This helps detect when a function that should be static is accidentally declared extern.

Most of the changes are to add missing header files so functions can see their declarations.

In a some cases functions that should have been static were marked as such. There were only five of these in the core but every little bit counts.

Lastly, it was necessary to suppress the warning in the postgres test modules where the function declarations are not available. This is fixable by aligning the module with the auto-generated code in core, but is not a priority.