Enable redundant-decls compiler warning.
Warn if anything is declared more than once in the same scope, even when the extra declaration is valid and changes nothing. This is primarily useful for catching missing header ifdef barriers.

Move the environ variable into config/parse.h since it must be declared by us and we use it multiple times.