Make the tests from allTest more stable (#48)
In the next set of circumstances, the test could fail:

1. After restoring the Postgres instance, the recovery process starts. Which
calls the asynchronous archive-get.
2. After archive-get checks the existence of the queue directory, but before
writing the WAL file, there are restores in the next test is begun. Which leads
to the deletion of the queue directory.
3. Since the directory no longer exists, writing the WAL file will fail, and
archive-get will write the error file to the queue.
4. A new Postgres instance will start, and the recovery process will begin,
which requests the WAL file.
5. The new archive-get looks into the queue directory, finds the error file, and
throws out the error, after which the Postgres recovery fails. Because the
previous archive-get background process has not finished yet.

This patch fixes the problem by using a separate spool directory for each test.

Ticket: ADBDEV-6384