Always restore system objects of database (#49)
Previously, when restoring with an empty filter (an empty array), the system
catalog was restored only for these databases: postgres, template1 and
template0. In this case, we cannot get metadata from user databases. Because of
this, it was required to get the OIDs of the user databases first and restore
again with user databases, whose OIDs had been obtained on the previous step. To
do one step to get metadata about tables instead of these two steps, the system
catalog is restored for all restoring databases.