Implement set option for pgbackrest verify (#51)
Implement set option for pgbackrest verify

This patch adds support for `--set` option for pgbackrest verify command. This
option was already previously documented, but wasn't actually implemented.

Default behavior for pgbackrest verify is to check all the backups present.
With new `--set` option only the specified backup will be verified (similar to
how `--set` works for pgbackrest restore). If the specified backup label is
not valid an error is thrown.

Ticket: ADBM-1941