Add cast to suppress Coverity complaint about volatile used in assert().
Coverity had this complaint:

assert_side_effect: Argument openData of ASSERT() has a side effect because the variable is volatile. The containing function might work differently in a non-debug build.

It appears this can also be fixed by assigning the volatile variable to an automatic but the cast seems to work just as well.