Implement restore progress for pgbackrest
The current implementation of the restore command did not provide any progress information. This patch implements it similarly to the backup command and also updates the info command to display restore progress alongside backup progress.

Restore is a read operation and should not block other commands. The only exception is that multiple restores with the same lock and repo path are not allowed, as each restore must write progress to a separate file. Therefore, no lock is needed for restores with a remote role and restore command should not be terminated when the stop command is executed.

Co-authored-by: Maxim Michkov <m.michkov@arenadata.io>