Fix the processing of multi-page WAL records at the beginning and end of the WAL file (#70)
WAL records can be split into 2 files. If such a record is located at the
beginning of the current file, then we need to first read the beginning of this
record from the previous file. Depending on how much of the record is in the
previous file, we either filter the records or write what is in the current file
as it is. The second case also occurs if the previous file does not exist in the
archive. Previously, the xlp_rem_len field was used to determine how much data
needs to be copied as is from the beginning of the file. The problem is that
this field does not take into account the size of the headers that are inside
the record if its size is greater than one page. Because of this, the record was
not fully copied, which led to further errors in reading the WAL. The issue
could also occur when reading the previous file, where there may also be an
unfinished record at the beginning that needs to be skipped.
A similar situation may occur when writing the last record in a segment file.
The part of the record that we read was copied to the end of the file, without
taking into account that more than one page could have been read. Therefore, in
addition to the record itself, it is also necessary to write the headers.

Fix reading at the beginning of the file by adding a loop to the copying process
in which we copy data sequentially from each of the pages, taking into account
their headers.
Fix the writing at the end of the file by replacing simply copying with using
the WriteRecord function, which writes taking into account the headers. Also add
a check for the end of the segment file to this function.
Also add various tests for multi-page records when reading from a previous file.

(cherry picked from commit 00c4fad4e221ee3ee85886ac42a05bc61f41bf42)