Allow verification of a specified backup.
Add support for the verify command --set option. This (internal) option was already accepted without errors but was not implemented.

The default behavior for verify is to check all the backups present. With the --set option only the specified backup will be verified. If the specified backup label is not valid an error is added to the result and verification is skipped. In addition, only WAL required to make the specified backup consistent will be verified.