Add LSTDEF() macro.
This macro allows static List objects to be created which simplifies usage when passing lists to functions. Also, since List objects are commonly used this makes the code base a bit more consistent.

For now skip static lists that are only used locally within a function since the benefit is not as clear.