Enable format-overflow=2 compiler warning.
Warn about calls to formatted input/output functions such as sprintf and vsprintf that might overflow the destination buffer.

This warning found a few short float buffers. In practice these are unlikely to ever be a problem based on our float usage but best to be safe.

Note that this warning only requires the buffer size to be 317 but it must be 318 to prevent warnings from -Wformat-truncation=2. We are not ready to enable that warning yet but seems better to get the buffer correct now.