Update invalid address check in common/io-tls unit test.
The prior check would result in a segfault from musl libc, which is understandable but not OK for a test.

Instead set all the flags to get the expected invalid result.

The second invalid test was not required for coverage so remove it. This was likely added before the earlier test was required for additional coverage.