Remove cfgParseOptionDefault() and cfgParseOptionRequired().
It is currently not possible to determine the default of all options knowing just the command. Some defaults are set in cfgLoadUpdateOption() and in an upcoming commit defaults may be based on the value of other options.

It would be possible to update parser to provide this information but that will complicate the parser and since the logic is only used to simplify options passed to remotes it does not seem worth the effort.

For the same reason cfgParseOptionRequired() can also be removed.