Fix wal filtering when there is no unfinished record at the end of the file (#76)
When there is no unfinished record at the end of the WAL segment, the next
record will be read, but when the next page is requested, it will be found that
the segment file has ended. However, since the current step has already been set
to stepBeginOfRecord, it was mistakenly assumed that there is an unfinished
record at the end that needs to be written. And since this is also the beginning
of the page, an attempt will be made to write down its header, which is missing,
which led to a crash.

Fix this by improving the condition for detecting the presence of an unfinished
record at the end of the file. Now it is assumed that the current record is
incomplete only if we have at least started reading the header. A test has also
been added for another edge case if the WAL switch record is located at the very
end of the page.

(cherry picked from commit e0662bb152e649984e38e57e0d216da49a776d8b)