Remove record is too big error (#80)
This error meant that the input buffer was not enough to read the unfinished
record at the beginning of the file. It was considered an unlikely situation.
But practice has shown that this situation is possible.

Fix it by getting rid of this error. Now, if there is not enough input buffer,
we can request the next buffer. Also, the logic of reading an unfinished record
at the beginning of a file, if there is a previous file, is unified with the
general logic of reading records.

Ticket: ADBM-2453
(cherry picked from commit 1d9c425cf27c02731f2ba86c222f5ddb8267f135)