Implement JSON output in verify command (#82)
Implement JSON output in verify command

pgbackrest verify command result code is hard to rely on when it is required to
understand the correctness of the backup data in the repository. It is possible
to get more information by adding options --output text --verbose,

But parsing such output is hard and error prone. This patch implements
json format as a new output format for verify command which transforms the
output into json.