Replace the error of missing the next WAL file with a warning (#83)
In the case when the next file is on a different timeline, we could fall with
the error "The file with the end of the LSN record is missing", but in this
case, the absence of the next file is not an error. Because the unfinished
record at the end of the current file will be overwritten at the beginning of
the next one. If the next file is really not in the archive, gpdb will catch
this error when trying to request that file.

Fix this by replacing the error with a warning and writing the unfinished record
as is.
The case of missing the previous file is already being handled, so just add a
warning.

(cherry picked from commit 14dd42c4785b077533b6ae74a5c352db57ddb904)