Fix reading the end of a record when the next file is partial (#85)
When there is an unfinished record at the end of the file, the next file may be
partial if the timeline was switched during its writing. Then the end of the
record may not be fully written. In this case, we will start reading garbage at
some point, which will lead to an error.

Fix this by wrapping the getEndOfRecord call in try/catch and replacing the
error with a warning and writing the record as is.

Ticket: ADBM-2468
(cherry picked from commit e8e84c249734cd6981c42a46af3b08f0d3240733)