Fix incorrect pg_control struct used for PostgreSQL 9.5.
The 9.6 pg_control struct was being used for 9.5. This was not detected by testing because the new field introduced for 9.6 fit into an alignment hole in the 9.5 struct so the size of the struct and offset of all other members did not change. Since the new member was not used there was no impact on functionality.

9.5 is being removed in the next release so the only reason to fix this is to make the diff for that change more sensible, and to document that this happened.