Migrate document command/option description preprocessing to C.
The Perl processing of config.yaml put hard limits on the format of that file. To allow flexibility in the file format remove all Perl processing on config.yaml.

This is just the beginning of migrating the preprocessor to C but even this small bit allows the removal of a lot of Perl code.