Show internal state for commands/options in parse.auto.c.inc.
Previously internal state was not included because it does not affect how commands/options are used -- only whether they are shown in help. However, this makes it hard to know when the internal state changes because help is generated at build time and in any case is just a binary blob.

Internal state is not stored in the config structures since the macros resolve to nothing but it is handy for debugging to see when internal state has changed.