Use roles to simplify option command lists.
In many cases the valid commands are based on the commands valid for roles. In these cases derive the commands from a role list rather than an explicit command list.

Not only is this notation more compact but it helps prevent new commands from being missed.

This exposed a few issues:

1) The cmd option should only be valid when a command supports the local role since it is used to execute the local process. A number of commands were included before that did not have the local role.

2) cmd-ssh should be valid for any command that allows remotes. The annotate command was missing from this list.

3) compress-level-network should be valid for any command that allows remotes. The repo-rm command was missing from this list.